+++ flog/package.json
@@ -22,6 +22,7 @@
     "date-fns": "^2.28.0",
     "dompurify": "^2.3.6",
     "graphql": "^16.3.0",
+    "lodash": "^4.17.21",
     "next": "12.1.0",
     "react": "17.0.2",
     "react-beautiful-dnd": "^13.1.0",
@@ -38,6 +39,7 @@
     "@graphql-codegen/cli": "^2.6.2",
     "@types/apollo-upload-client": "^17.0.0",
     "@types/dompurify": "^2.3.3",
+    "@types/lodash": "^4.14.181",
     "@types/node": "17.0.21",
     "@types/react": "17.0.40",
     "@types/react-beautiful-dnd": "^13.1.2",
