+++ flog/pages/myTrips/index.tsx
@@ -1,5 +1,5 @@
 import MyTrip from "../../src/components/units/myTrip/MyTrip.container";
 
 export default function MyTripPage() {
-  return <MyTrip />;
+  return <MyTrip isMine={true} />;
 }
