+++ flog/pages/ourTrips/index.tsx
@@ -1,5 +1,5 @@
 import OurTrip from "../../src/components/units/ourTrip/OurTrip.container";
 
 export default function OurTripPage() {
-  return <OurTrip />;
+  return <OurTrip isMine={false} />;
 }
