+++ flog/pages/ourTrips/titleSearch/index.tsx
@@ -6,7 +6,6 @@ import OurTripBanner from "../../../src/components/units/ourTrip/banner/OurTripB
 import { FETCH_TITLE_SEARCH } from "../../../src/components/units/ourTrip/OurTrip.queries";
 import TitleSearchBanner from "../../../src/components/units/ourTrip/titleSearch/banner/TitleSearchBanner.container";
 import TitleSearchList from "../../../src/components/units/ourTrip/titleSearch/list/TitleSearchList.container";
-import TripList from "../../../src/components/units/tripList/TripList.container";
 
 const BodyContainer = styled.div`
   width: 100%;
@@ -16,20 +15,24 @@ const BodyContainer = styled.div`
 `;
 
 export default function TitleSearchPage() {
+  const [inputs, setInputs] = useState({ doName: "", cityName: "", title: "" });
+
   const { data: titleData } = useQuery(FETCH_TITLE_SEARCH, {
     variables: {
-      where: "강원도 횡성군",
-      search: "두번째",
+      where: `${inputs.doName}.${inputs.cityName}`,
+      search: inputs.title,
     },
   });
 
+  console.log(titleData);
+
   // 상위 컴포넌트에 넣을 내용 - MapModal
 
-  const [inputs, setInputs] = useState({ doName: "", cityName: "" });
   const [mapModal, setMapModal] = useState(false);
 
   const onClickMapModal = () => {
     setMapModal(true);
+    setInputs({ doName: "", cityName: "" });
   };
 
   const onClickExitMapModal = () => {
@@ -39,7 +42,6 @@ export default function TitleSearchPage() {
 
   const onClickSubmitMapModal = () => {
     setMapModal(false);
-    console.log(inputs.doName, inputs.cityName);
   };
 
   return (
@@ -58,6 +60,7 @@ export default function TitleSearchPage() {
           onClickExit={onClickExitMapModal}
           onClickSubmit={onClickSubmitMapModal}
           inputs={inputs}
+          setInputs={setInputs}
         />
         <TitleSearchList isMine={false} titleData={titleData} />
       </BodyContainer>
