+++ flog/src/components/units/myTrip/list/MyTripList.styles.ts
@@ -1,6 +1,7 @@
 import styled from "@emotion/styled";
+import { breakPoints } from "../../../../commons/styles/Media";
 
-export const List = styled.div`
+export const ListWrap = styled.div`
   width: 95%;
   max-width: 1200px;
   display: flex;
@@ -8,11 +9,32 @@ export const List = styled.div`
   align-items: center;
 `;
 
+export const List = styled.div`
+  padding-top: 80px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  @media ${breakPoints.mobile} {
+    padding-top: 30px;
+  }
+  @media ${breakPoints.tablet} {
+  }
+`;
+
 export const TripList = styled.div`
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   place-items: center;
   grid-gap: 30px;
+  @media ${breakPoints.mobile} {
+    grid-template-columns: repeat(2, 1fr);
+    grid-gap: 8px;
+    padding-left: 8px;
+    padding-right: 8px;
+  }
+  @media ${breakPoints.tablet} {
+  }
 `;
 
 export const Head = styled.div`
@@ -20,36 +42,24 @@ export const Head = styled.div`
   display: flex;
   justify-content: space-between;
   margin-bottom: 70px;
+  @media ${breakPoints.mobile} {
+    margin-bottom: 40px;
+  }
+  @media ${breakPoints.tablet} {
+  }
 `;
 
 export const Label = styled.div`
   color: #2c3131;
   font-weight: 500;
   font-size: 2rem;
-`;
-
-export const Select = styled.select`
-  width: 154px;
-  height: 56px;
-  border: 1px solid #cacaca;
-  border-radius: 8px;
-  padding-left: 24px;
-  color: #626262;
-  font-size: 1rem;
-  line-height: 150%;
-  background: url("/img/icon-mytrip-write-card-lower.svg") no-repeat 90% 50%;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-  select::-ms-expand {
-    display: none;
+  @media ${breakPoints.mobile} {
+    font-weight: 400;
+    font-size: 20px;
   }
-  :hover {
-    cursor: pointer;
+  @media ${breakPoints.tablet} {
   }
 `;
-
-export const SelectOption = styled.option``;
 
 export const CardWrapper = styled.div`
   width: 100%;
@@ -67,8 +77,13 @@ export const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: start;
-  cursor: pointer;
   position: relative;
+  @media ${breakPoints.mobile} {
+    height: 240px;
+    border-radius: 10px;
+  }
+  @media ${breakPoints.tablet} {
+  }
 `;
 
 export const Mark = styled.div`
@@ -77,6 +92,15 @@ export const Mark = styled.div`
   img {
     position: relative;
     left: 32px;
+  }
+  @media ${breakPoints.mobile} {
+    img {
+      position: relative;
+      width: 17px;
+      left: 18px;
+    }
+  }
+  @media ${breakPoints.tablet} {
   }
 `;
 
@@ -88,6 +112,18 @@ export const EditWrap = styled.div`
     position: relative;
     top: 21px;
     width: 1.5rem;
+    :hover {
+      cursor: pointer;
+    }
+  }
+  @media ${breakPoints.mobile} {
+    right: 12px;
+    img {
+      width: 17px;
+      top: 10px;
+    }
+  }
+  @media ${breakPoints.tablet} {
   }
 `;
 
@@ -99,7 +135,12 @@ export const Image = styled.div`
   border-radius: 20px;
   z-index: 0;
   aspect-ratio: 38 / 32.8;
-
+  cursor: pointer;
+  @media ${breakPoints.mobile} {
+    border-radius: 10px;
+  }
+  @media ${breakPoints.tablet} {
+  }
   :hover {
     img {
       width: 100%;
@@ -125,38 +166,72 @@ export const Text = styled.div`
   display: flex;
   flex-direction: column;
   text-align: left;
-  margin-top: 1.2rem;
+  padding-right: 5px;
+  padding-left: 5px;
 `;
 
 export const Title = styled.div`
   font-size: 1.5rem;
-  line-height: 1.5rem;
-  text-align: left;
-  margin-bottom: 0.5rem;
-  line-height: 120%;
+  line-height: 170%;
+  padding-top: 20px;
+  word-break: keep-all;
+  @media ${breakPoints.mobile} {
+    font-weight: 300;
+    font-size: 16px;
+    padding-top: 12px;
+    line-height: 130%;
+  }
+  @media ${breakPoints.tablet} {
+  }
+`;
+
+export const Date = styled.li`
+  font-size: 20px;
+  color: #a9a9a9;
+`;
+
+export const Region = styled.li`
+  font-size: 20px;
+  color: #a9a9a9;
+  font-weight: 200;
 `;
 
 export const Subtitle = styled.div`
-  font-size: 1.1rem;
   display: flex;
-  flex-direction: row;
   line-height: 140%;
-`;
-
-export const Date = styled.div`
-  font-size: 1.1rem;
-  color: #a9a9a9;
-  margin-right: 6px;
-`;
-
-export const Region = styled.div`
-  font-size: 1.1rem;
-  color: #a9a9a9;
-  font-weight: 300;
-  line-height: 28px;
-`;
-
-export const NewList = styled.div`
-  margin-top: 60px;
-  cursor: pointer;
+  ul {
+    li {
+      list-style: none;
+      float: left;
+      color: #a9a9a9;
+    }
+    li::before {
+      content: "|";
+      font-weight: 200;
+      margin-left: 8px;
+      margin-right: 8px;
+    }
+    li:first-of-type::before {
+      content: "";
+      margin-left: 0px;
+      margin-right: 0px;
+    }
+  }
+  .location {
+    font-weight: 200;
+  }
+  @media ${breakPoints.mobile} {
+    ul {
+      li {
+        font-size: 12px;
+        line-height: 140%;
+      }
+      li::before {
+        margin-left: 0px;
+        margin-right: 4px;
+      }
+    }
+  }
+  @media ${breakPoints.tablet} {
+  }
 `;
