+++ flog/src/components/units/myTrip/myTrip.container.tsx
@@ -5,6 +5,10 @@ import NewTripScheduleModal from "../../commons/modals/newSchedule/mainSchedule/
 import MyTripBanner from "./banner/MyTripBanner.container";
 import MyTripList from "./list/MyTripList.container";
 import { FETCH_SCHEDULES, FETCH_USER } from "./MyTrip.queries";
+import * as Non from "./MyTrip.styles";
+import OutlinedButton01 from "../../commons/buttons/outlined/01/OutlinedButton01.container";
+import { v4 as uuid4 } from "uuid";
+import { useRouter } from "next/router";
 
 const BodyContainer = styled.div`
   width: 100%;
@@ -13,13 +17,13 @@ const BodyContainer = styled.div`
   align-items: center;
 `;
 
-export default function MyTrip() {
-  const { data: myData, fetchMore } = useQuery(FETCH_SCHEDULES, {
+export default function MyTrip(props) {
+  const router = useRouter();
+  const { data: myData } = useQuery(FETCH_SCHEDULES, {
     variables: { page: 1 },
   });
 
   const { data: userData } = useQuery(FETCH_USER);
-  // console.log(userData?.fetchUser);
 
   const [newScheduleModal, setNewScheduleModal] = useState(false);
 
@@ -27,25 +31,8 @@ export default function MyTrip() {
     setNewScheduleModal((prev) => !prev);
   };
 
-  const onClickMore = () => {
-    if (!myData) return;
-
-    fetchMore({
-      variables: {
-        page: Math.ceil(myData.fetchSchedules.length / 12) + 1,
-      },
-      updateQuery: (prev, { fetchMoreResult }) => {
-        if (!fetchMoreResult.fetchSchedules)
-          return { fetchSchedules: [...prev.fetchSchedules] };
-
-        return {
-          fetchSchedules: [
-            ...prev.fetchSchedules,
-            ...fetchMoreResult.fetchSchedules,
-          ],
-        };
-      },
-    });
+  const onClickMoveToOurTrip = () => {
+    router.push("/ourTrips");
   };
 
   return (
@@ -62,13 +49,38 @@ export default function MyTrip() {
           onClickMapModal={onClickNewScheduleModal}
           myData={myData}
         />
-        <MyTripList
-          isMine={true}
-          userData={userData}
-          myData={myData}
-          fetchMore={fetchMore}
-          onClickMore={onClickMore}
-        />
+        {myData?.fetchSchedules.length > 0 ? (
+          <MyTripList />
+        ) : (
+          <Non.ListWrap>
+            <Non.List>
+              <Non.Text>
+                지금까지{" "}
+                <span style={{ fontWeight: 700, color: "#58BD97" }}>
+                  16,930개
+                </span>
+                의 여행이
+                <br />
+                기록되었어요
+              </Non.Text>
+              <Non.TripList>
+                {[1, 2, 3, 4, 5, 6].map((el) => (
+                  <Non.Image key={uuid4()}>
+                    <img
+                      src={`/img/trips/notUser/trip-img${el}.png`}
+                      alt="여행이미지"
+                    />
+                  </Non.Image>
+                ))}
+              </Non.TripList>
+              <OutlinedButton01
+                content="우리의 여행 보러가기"
+                size="large"
+                onClick={onClickMoveToOurTrip}
+              />
+            </Non.List>
+          </Non.ListWrap>
+        )}
       </BodyContainer>
     </>
   );
