+++ flog/src/components/units/ourTrip/OurTrip.container.tsx
@@ -3,12 +3,11 @@ import { useState } from "react";
 import MapModal from "../../commons/modals/map/MapModal.container";
 import TripList from "../tripList/TripList.container";
 import OurTripBanner from "./banner/OurTripBanner.container";
+import OurTripList from "./list/OurTripList.container";
 import { FETCH_SHARE_SCHEDULES } from "./OurTrip.queries";
 import { BodyContainer } from "./OurTrip.styles";
 
-export default function OurTrip() {
-  const { data: shareData, fetchMore } = useQuery(FETCH_SHARE_SCHEDULES);
-
+export default function OurTrip(props) {
   // 상위 컴포넌트에 넣을 내용 - MapModal
 
   const [inputs, setInputs] = useState({ doName: "", cityName: "" });
@@ -28,29 +27,6 @@ export default function OurTrip() {
     console.log(inputs.doName, inputs.cityName);
   };
 
-  const onClickMore = () => {
-    // alert("더보기 버튼");
-    //  console.log(shareData);
-    if (!shareData) return;
-
-    fetchMore({
-      variables: {
-        page: Math.ceil(shareData.fetchShareSchedules.length / 12) + 1,
-      },
-      updateQuery: (prev, { fetchMoreResult }) => {
-        if (!fetchMoreResult.fetchShareSchedules)
-          return { fetchShareSchedules: [...prev.fetchShareSchedules] };
-
-        return {
-          fetchShareSchedules: [
-            ...prev.fetchShareSchedules,
-            ...fetchMoreResult.fetchShareSchedules,
-          ],
-        };
-      },
-    });
-  };
-
   return (
     <>
       {mapModal && (
@@ -68,11 +44,7 @@ export default function OurTrip() {
           onClickSubmit={onClickSubmitMapModal}
           inputs={inputs}
         />
-        <TripList
-          isMine={false}
-          shareData={shareData}
-          onClickMore={onClickMore}
-        />
+        <OurTripList />
       </BodyContainer>
     </>
   );
